Update mksh to the released version R40 (no code changes)
Adjust build system to changes in both Android and mksh
and make further updates and test builds easier

+++ mksh/mkmf.sh
@@ -1,4 +1,4 @@
-# Copyright © 2010
+# Copyright © 2010, 2011
 #	Thorsten Glaser <t.glaser@tarent.de>
 # This file is provided under the same terms as mksh.
 #-
@@ -7,6 +7,16 @@
 #
 # This script is supposed to be run from/inside AOSP by the
 # porter of mksh to Android (and only manually).
+
+if test x"$1" = x"-t"; then
+	# test compilation
+	args=-r
+	mkmfmode=1
+else
+	# prepare for AOSP
+	args=-M
+	mkmfmode=0
+fi
 
 cd "$(dirname "$0")"
 srcdir=$(pwd)
@@ -45,13 +55,12 @@ LIBS=
 CC=$aospdir/prebuilt/linux-x86/toolchain/arm-eabi-4.4.0/bin/arm-eabi-gcc
 addvar CPPFLAGS -I$aospdir/system/core/include \
     -I$aospdir/hardware/libhardware/include \
-    -I$aospdir/system/core/include \
-    -I$aospdir/hardware/libhardware/include \
     -I$aospdir/hardware/libhardware_legacy/include \
     -I$aospdir/hardware/ril/include \
     -I$aospdir/dalvik/libnativehelper/include \
     -I$aospdir/frameworks/base/include \
     -I$aospdir/frameworks/base/opengl/include \
+    -I$aospdir/frameworks/base/native/include \
     -I$aospdir/external/skia/include \
     -I$aospdir/out/target/product/generic/obj/include \
     -I$aospdir/bionic/libc/arch-arm/include \
@@ -63,25 +72,34 @@ addvar CPPFLAGS -I$aospdir/system/core/include \
     -I$aospdir/bionic/libm/include/arch/arm \
     -I$aospdir/bionic/libthread_db/include \
     -D__ARM_ARCH_5__ -D__ARM_ARCH_5T__ -D__ARM_ARCH_5E__ -D__ARM_ARCH_5TE__ \
-    -I$aospdir/system/core/include/arch/linux-arm/ \
     -include $aospdir/system/core/include/arch/linux-arm/AndroidConfig.h \
+    -I$aospdir/system/core/include/arch/linux-arm/ \
     -DANDROID -DNDEBUG -UDEBUG
 addvar CFLAGS -fno-exceptions -Wno-multichar -msoft-float -fpic \
-    -ffunction-sections -funwind-tables -fstack-protector -fno-short-enums \
-    -march=armv5te -mtune=xscale -mthumb-interwork -fmessage-length=0 \
-    -W -Wall -Wno-unused -Winit-self -Wpointer-arith -Werror=return-type \
-    -Werror=non-virtual-dtor -Werror=address -Werror=sequence-point \
-    -Wstrict-aliasing=2 -finline-functions -fno-inline-functions-called-once \
-    -fgcse-after-reload -frerun-cse-after-loop -frename-registers -mthumb \
-    -Os -fomit-frame-pointer -fno-strict-aliasing -finline-limit=64
+    -ffunction-sections -funwind-tables -fstack-protector -Wa,--noexecstack \
+    -Werror=format-security -fno-short-enums -march=armv5te -mtune=xscale \
+    -Wno-psabi -mthumb-interwork -fmessage-length=0 -W -Wall -Wno-unused \
+    -Winit-self -Wpointer-arith -Werror=return-type -Werror=non-virtual-dtor \
+    -Werror=address -Werror=sequence-point -Wstrict-aliasing=2 \
+    -finline-functions -fno-inline-functions-called-once -fgcse-after-reload \
+    -frerun-cse-after-loop -frename-registers -mthumb -Os \
+    -fomit-frame-pointer -fno-strict-aliasing -finline-limit=64
 addvar LDFLAGS -nostdlib -Bdynamic -Wl,-T,$aospdir/build/core/armelf.x \
     -Wl,-dynamic-linker,/system/bin/linker -Wl,--gc-sections \
-    -Wl,-z,nocopyreloc -Wl,--no-undefined \
+    -Wl,-z,nocopyreloc -Wl,-z,noexecstack -Wl,--no-undefined \
     $aospdir/out/target/product/generic/obj/lib/crtbegin_dynamic.o
 addvar LIBS -L$aospdir/out/target/product/generic/obj/lib \
     -Wl,-rpath-link=$aospdir/out/target/product/generic/obj/lib -lc \
     $aospdir/prebuilt/linux-x86/toolchain/arm-eabi-4.4.0/bin/../lib/gcc/arm-eabi/4.4.0/interwork/libgcc.a \
     $aospdir/out/target/product/generic/obj/lib/crtend_android.o
+
+
+### Flags used by test builds
+if test $mkmfmode = 1; then
+	addvar CPPFLAGS '-DMKSH_DEFAULT_EXECSHELL=\"/system/bin/sh\"'
+	addvar CPPFLAGS '-DMKSH_DEFAULT_TMPDIR=\"/sqlite_stmt_journals\"'
+	addvar CPPFLAGS '-DMKSHRC_PATH=\"/system/etc/mkshrc\"'
+fi
 
 
 ### Override flags
@@ -95,7 +113,7 @@ addvar CPPFLAGS -DMKSH_NOPWNAM
 #addvar CPPFLAGS -DMKSH_NO_LIMITS
 
 # Set target platform
-TARGET_OS=Linux
+TARGET_OS=Android
 # Building with -std=c99 or -std=gnu99 clashes with Bionic headers
 HAVE_CAN_STDG99=0
 HAVE_CAN_STDC99=0
@@ -108,16 +126,14 @@ HAVE_CAN_FSTACKPROTECTORALL=0
 HAVE_CAN_WALL=0
 export HAVE_CAN_FNOSTRICTALIASING HAVE_CAN_FSTACKPROTECTORALL HAVE_CAN_WALL
 
-# disable the mknod(8) built-in to get rid of needing setmode.c
-HAVE_MKNOD=0; export HAVE_MKNOD
-
 # even the idea of persistent history on a phone is funny
 HAVE_PERSISTENT_HISTORY=0; export HAVE_PERSISTENT_HISTORY
 
 # ... and run it!
 export CC CPPFLAGS CFLAGS LDFLAGS LIBS TARGET_OS
-sh ../src/Build.sh -M
+sh ../src/Build.sh $args
 rv=$?
+test x"$args" = x"-r" && exit $rv
 test x0 = x"$rv" && mv -f Makefrag.inc ../
 cd ..
 rm -rf tmp
